separated histogram-creating and histogram statistics functions into separate files
