+++ src/dspeed/processors/__init__.py
@@ -79,7 +79,7 @@ _modules = {
     "get_wf_centroid": "get_wf_centroid",
     "histogram": "histogram",
     "histogram_around_mode": "histogram",
-    "histogram_stats": "histogram",
+    "histogram_stats": "histogram_stats",
     "inject_damped_oscillation": "inject_ringing",
     "inl_correction": "inl_correction",
     "moving_slope": "kernels",
